refactor(5.4-6.0): add mapping for XP_Plan.aendert and wurdeGeandertVon
The attributes 'XP_Plan.aendert' and 'XP_Plan.wurdeGeandertVon' are renamed
with XPlanVersion 6.0 to 'XP_Plan.aendertPlan' and 'XP_Plan.wurdeGeandertVonPlan',
so this mapping is needed.
Also the the attribute 'rechtscharakter' is renamed to 'aenderungsart' and
the values are adjusted with XPlanVersion 6.0. . This is considered here with
a classification.

SVC-1179