fix(5.4-6.0): delete mapping of aendert and wurdeGeandertVon
Delete the mapping of child properties of 'aendert' and 'wurdeGeandertVon' for
BP_Plan, FP_Plan, LP_Plan, RP_Plan and SO_Plan if the mapping exists. This
mapping was incomplete, as it wasn't set up for each listed type.
The mapping is also not needed anymore for each type because the mapping is set
up once for parent type XP_Plan.
If a structural retype is set up for the listed type the attributes 'aendert'
and 'wurdeGeandertVon' are excluded (if not already done).

SVC-1179