refactor(5.4-6.0): add mapping of BP/FP/.._TextAbschnitt.rechtscharakter
Add classification for the attribute 'rechtscharakter' within the retype of
BP_TextAbschnitt/FP_../RP_../LP_../ and SO_TextAbschnitt to XP_TextAbschnitt.
This is needed because the previously used types of 'rechtscharakter'
BP/FP/.._Rechtschrakter changed to XP_Rechtscharakter with new enumeration values.

Also the attribute 'rechtschrakter' is excluded in the respective structural retypes.

SVC-1179