refactor(5.4-6.0): add mapping of BP_Plan.verfahren
The attribute 'BP_Plan.verfahren' on source side needs to be mapped to
'BP_Bereich.verfahren' on the target side as this changes with XPlan
version 6.0.
To archiv this a groovy join is set up which collects the values of
'BP_Plan.verfahren' for each 'BP_Bereich.id'. The collector is than used
to assign the values to 'BP_Bereich.verfahren' on target side within the
existing rename between BP_Bereich on target and source side.

SVC-1179