refactor(5.4-6.0): add mapping of attributes in BP_BaugebietsTeilFlaeche
With XPlan 6.0 the attribute 'sondernutzung' in BP_BaugebietsTeilFlaeche is
replaced with an complex type. So the attributes 'sondernutzung',
'detaillierteSondernutzung' and 'nutzungText' on the souzce side need to be
mapped to attributes of the complex type 'sondernutzung' on the target side.
Also the relevant attributes are excluded from the structural retype of
BP_BaugebietsTeilFlaeche.

SVC-1179