refactor(5.4-6.0): add mapping BP_GemeinschaftsanlagenFlaeche
With XPlan 6.0 the attribute 'zweckbestimmung' in BP_GemeinschaftsanlagenFlaeche
is replaced with an complex type. So the attributes 'zweckbestimmung' and
'detaillierteZweckbestimmung' on the source side need to be mapped to attributes
of the complex type 'zweckbestimmung' on the target side.
Also the relevant attributes are excluded from the structural retype of
BP_GemeinschaftsanlagenFlaeche.

SVC-1179