fix(5.4-6.0): add mapping of detaillierteZweckbestimmung
With XPlan 6.0 the attribute 'detaillierteZweckbestimmung' in BP_WaldFlaeche,
BP_LandwirtschaftsFlaeche, BP_Landwirtschaft, BP_GruenFlaeche and
BP_VerEntsorgung is moved to the attribute 'detail' of the complex type
'zweckbestimmung'. That was missed in the mapping and is added here.
Also the attribute 'detaillierteZweckbestimmung' is excluded from the
respective structural retypes.

SVC-1179