refactor(5.4-6.0): add mapping for BP_VerkehrsflaecheBesondererZweckbestimmung
The feature type BP_VerkehrsflaecheBesondererZweckbestimmung is mapped to SO_Strassenverkehr.
With this change the attributes 'zweckbestimmung' and 'detaillierteZweckbestimmung' on source
side need to be mapped to the attributes artDerFestlegung.allgemein and artDerFestlegung.detail
on target side.
As the values of the old enumeration 'BP_ZweckbestimmungStrassenverkehr' of zweckbestimmung on
source side changes to new values of the enumeration 'SO_ZweckbestimmungStrassenverkehr' used
for 'allgemein' a classification is needed.
No classification set set up ewuivalent for 'BP_DetailZweckbestStrassenverkehr' which changes to
'SO_DetailZweckbestStrassenverkehr' because these enumerations contain no values.

Both attributes are also excluded in the structural retype of
BP_VerkehrsflaecheBesondererZweckbestimmung to SO_Strassenverkehr.

SVC-1179