refactor(5.4-6.0): add mapping of attributes of BP_GewaesserFlaeche
The feature type 'BP_GewaesserFlaeche' is mapped to 'SO_Gewaesser'. The attributes
'zweckbestimmung' and 'deatillierteZweckbestimmung' on source side need to be
mapped to 'artDerFestlegung.allgemein' and 'artDerFestlegung.detail' on the
target side. This is done here.
Also the relevant attributes are excluded from the structural retype of
'BP_GewaesserFlaeche'.

SVC-1179