refactor(5.4-6.0): add mapping detaillierteZweckbsetimmung for FP-types
The attribute 'detaillierteZweckbsetimmung' on source side need to be mapped to
'detail' the attribute of the new complex attribut 'zweckbestimmung' on target side.
This is done here for the feature types FP_Gemeinbedarf, FP_SpielSportanlage,
FP_WaldFlaeche, FP_Landwirtschaft, FP_LandwirtschaftsFlaeche, FP_Gruen and
FP_VerEntsorgung.
For FP_VerEntsorgung also 'textlicheErgaenzung' is mapped to 'textlicheErgänzung' of
the new complex attribute 'zweckbestimmung'.
Also the attributes are excluded from the respective structural retypes.

SVC-1179