refactor(5.4-6.0): add mapping FP_Strassenverkehr.zweckbestimmung
With XPlan 6.0 the feature type FP_Strassenverkehr need to be mapped to SO_Strassenverkehr.
And with this the attributes 'zweckbestimmung' and 'detaillierteZweckbestimmung' on source
side need to be mapped to the attributes 'allgemein' and 'detail' of the complex type
'artDerFestlegung' on target side. This is done here.
Also both attributes are excluded from structural retype, and 'rechtscharakter' is also
excluded as this was missed before.

SVC-1179