refactor(5.4-6.0): add mapping FP_Gewaesser.zweckbestimmung
With XPlan 6.0 the feature type FP_Gewaesser need to be mapped to SO_Gewaesser.
And with this the attributes 'zweckbestimmung' and 'detaillierteZweckbestimmung' on source
side need to be mapped to the attributes 'allgemein' and 'detail' of the complex type
'artDerFestlegung' on target side. This is done here.
Also both attributes are excluded from structural retype, and 'rechtscharakter' is also
excluded as this was missed before.
Same is done for FP_Wasserwirtschaft (mapped to SO_Wasserwirtschaft), but the attributes
on target side are 'artDerFestlegung' and 'detailArtDerFestlegung'.

SVC-1179