refactor(5.4-6.0): add XP_RasterDarstell.refScan to BP_Bereich.refScan
Previously, BP_Bereich on source side was mapped with a retype to BP_Bereich on target side.
This is replaced with a join of BP_Bereich and XP_RasterDarstellung to BP_Bereich. It is
needed because XP_RasterDatrstellung.refScan (linked via rasterBasis to BP_Bereich) is
mapped to BP_Bereich.refScan because the attribute rasterBasis is deleted with XPlan 6.0.
Also the mapping of all attributes of BP_Bereich are set up as the previously used structural
retype is replaced with a join.

SVC-1179