fix(5.4-6.0): XP_RasterDarstellung.refLegende to BP_Plan.externeRef
Previously, the XP_RasterDarstellung.refLegende.XP_ExterneReferenz was mapped
with an 'Inline-Transformation' to BP_Plan.externeReferenz.XP_SpezExterneReferenz.
This 'Inline-Transformation' is replaced with a 'Rename' functio. Using an 'Inline-
Transformatio' here is not correct as the mapping is between differenz feature types,
XP_ExterneReferenz and XP_SpezExterneReferenz.
Mapping of all subattributes of XP_RasterDarstellung.refLegende.XP_ExterneReferenz
are added here accordingly.

SVC-1179