refactor(5.4-6.0): add mapping of attribute of BP_WasserwirtschaftsFl
With XPlan 6.0 'BP_WasserwirtschaftsFlaeche' is replaced with 'SO_Wasserwirtschaft'.
And the the attribute 'detaillierteZweckbestimmung' on source side needs to be mapped
to 'detailArtDerFestlegung' on target side. Previously, this was missed and is added
here. The attribute is excluded from the structural retype.

Also the following attributes are excluded from its respective structural retypes as
this was missed previously: SO_Gewaesser.artDerFestlegung, SO_Gewaesser.detailArtDerFestlegung,
SO_Strassenverkehrsrecht.artDerFestlegung, SO_Strassenverkehrsrecht.detailArtDerFestlegung.

SVC-1179