refactor(5.4-6.0): add classifications to SO_Objekt.rechtscharakter
Some BP_Objekte are mapped to SO_Objekte and FP_Objekte to SO_Objekte (e.g.
BP_StrassenVerkehrsFlaeche to SO_Strassenverkehr or BP_GewaesserFlaeche to
SO_Gewaesser). These retypes do not inherit the respective calssification of
the attribute 'rechtscharacter' which is defined for the abstract mapping
between BP_Objekt to BP_Objekt and FP_Objekt to FP_Objekt.
Therefore a classification is needed for
BP_Objekt.rechtschrakter to SO_Objekt.rechtscharakter and
FP_Objekt.rechtschrakter to SO_Objekt.rechtscharakter. This is added here.

SVC-1179