fix(5.4-6.0): split mapping of SO_Luftverkehrsrecht
Previously, the feature type 'SO_Luftverkehrsrecht' on source side was mapped to the same
feature type on target side and a classification for the attribute 'artDerFestlegung' was
set up. But this classification was incorrect and is deleted here.
The correct mapping is: If the attribute 'artDerFestlegung' is 7000 'SO_Luftverkehrsrecht'
should be mapped to 'SO_Baubeschraenkung' with 'artDerFestlegung' = 2000. This is set up
here.
If 'artDerFestlegung' is not 7000 'SO_Luftverkehrsrecht' is still mapped to
'SO_Luftverkehrsrecht'.

SVC-1179