refactor(5.4-6.0): split mapping of SO_SonstigesRecht
Previously, the feature type 'SO_SonstigesRecht' on source side was mapped to the same
feature type on target side. But if the attribute 'artDerFestlegung' is 1000 or 1200
'SO_SonstigesRecht' should be mapped to 'SO_Baubeschraenkung'. This is set up
here. Also classification for 'artDerFestlegung' is created.
If the value of 'artDerFestlegung' is not one of the mentioned values 'SO_SonstigesRecht'
is still mapped to 'SO_SonstigesRecht'.

SVC-1179