fix(5.4-6.0): fix mapping of XP_Rasterdarstellung.refText
Previously, the mapping of XP_Rasterdarstellung.refText to the attribute texte
(which links to XP_TextAbschnitt) for BP_Plan, FP_Plan and RP_Plan wasn't working
correctly. This is fixed here:
Mapping of XP_Rasterdarstellung.id was only set up for BP_Plan.texte.href and missed
in the mapping to FP_Plan and RP_Plan. This is changed here and the mapping takes
place on abstract level (to XP_Plan.texte.href).
The order of feature types in the used groovy join (FP_Bereich, FP_Plan and
XP_Rasterdarstellung to XP_TextAbschnitt) is changed so that XP_Rasterdarstellung
is on top (same for RP_Plan).
And the groovy joins are adjusted so that a target (XP_TextAbschnitt) is only created
if a joined FP/BP/RP_Plan exists.

SVC-1179