refactor (5.4-6.0): add inline transformation for XP_ExterneReferenz
For all attributes which contains a XP_ExterneReferenz or XP_SpezExterneReferenz
an inline transformation is set up (if not set up yet). And the respective attribute
is excluded from structural retype if needed.
This is set up for the following attributes:

XP_Plan.externeReferenz
XP_BegruendungAbschnitt.refText
refGebaeudequerschnitt in BP_BaugebietsTeilFlaeche and BP_UeberbaubareGrundstuecksFlaeche
and BP_WohngebaeudeFlaeche
refMassnahmenText and refLandschaftsplan in BP_AusgleichsFlaeche and BP_AusgleichsMassnahme
and BP_SchutzPflegeEntwicklungsFlaeche and BP_SchutzPflegeEntwicklungsMassnahme
and FP_AusgleichsFlaeche
BP_Veraenderungssperre.refBeschluss

SVC-1179