fix(5.4-6.0): correct mapping of attributes in FP/BP_Bereich
With XPlan 6.0 the attributes 'versionBauGBDatum/Text', 'versionBauNVODatum/Text' and
'versionSonstRechtsgrundlagenDatum/Text' of FP_Bereich and BP_Bereich are deleted.
These attributes should be mapped to the respective attributes in FP_Plan/BP_Plan,
but only if no attributes are already contained in the same attributes in FP_Plan/BP_Plan.
Previously, the mapping of some of these attributes were already set up, the remaining
mappings are added here.

SVC-1179