fix(5.4-6.0): fix mapping of FP_Strassenverkehr.zweckbestimmung
The feature type FP_Strassenverkehr is mapped to SO_Strassenverkehr and the attribute
'zweckbestimmung' is mapped to 'artDerFestlegung.allgemein'. But with this some of the
values of 'zweckbestimmung' need to be mapped to new values in 'artDerFestlegung'.
This is fixed here.
Also if 'zweckbestimmung' = 1300 the attribute 'istOrtsdurchfahrt' on target side
should be set to 'true'.

SVC-1179