fix: make sure that reference is only created when existing
In case the attribute refText is populated, an XP_Rasterdarstellung object
is resulting in an XP_TextAbschnitt object. The reference from XP_Plan to this
XP_TextAbschnitt object had been created in any case (independent from if the
refText attribute is populated or not). This is now fixed so that the reference
is only created if there is also a respective XP_TextAbschnitt.

SVC-1179