fix(4.1-5.4): correct mapping of geometries
Previously, if the source data contains geometries with curves and arcs this
leads to an error when exporting the transformed data. To avoid this the
parameter 'Copy geometry objects in structural rename' is added and set to 'false'
in all structural renames. With this option the original XML structure with
arcs and curves will be recreated.

WGS-2507