fix(4.1-5.4): fix mapping of refTextInhalt in Join
The feature types 'BP_Baugebiet' and 'BP_BaugebietsTeilFlaeche' on source side
are mapped via Join to the target feature 'BP_BaugebietsTeilFlaeche'.
Previously the mapping of the attribute 'refTextInhalt.href' was inherited
from 'BP_Objekt 'but this leads to an error when transforming the data.
This is changed here so that the inherited function is disabled and a axplicity
mappings for 'BP_BaugebietsTeilflaeche.refTextInhalt.href' and
'BP_Baugbiet.refTextInhalt.href' are added.

WGS-2507