fix: keep xml structure of geometries for transformed data
If source data elements contain geometries using arcs, those cannot be
exported from hale studio as it is not supporting arc geometries. When
setting the parameter "Copy geometry objects in structure rename" to
false, the XML strucutre of the geometries is kept and the transformed
data is therefore still containing the arc geometries.
For the migration alignment 4.1 to 5.4, this has already been changed in
[this PR](https://www.dev.diplanung.de/DefaultCollection/EfA%20DiPlanung/_git/hale-alignments/pullrequest/1693)

SVC-1678