fix(4.1-5.4): correct inherited mapping of 'informell' in join
The Join of 'BP_Baugebiet' and 'BP_BaugebietsTeilFlaeche' inherits the mapping
of sub-attributes of 'informell' to 'externeReferenz'.
But in case both source attributes are filled with values (in 'BP_Baugebiet' and in
'BP_BaugebietsTeilFlaeche') only one value is mapped.
This is changed here and the inherited functions are disabled and replaced with
explicit mappings to a separate instance context each.

WGS-2507