fix(5.4-6.0): add mapping of 'XP_AbstraktesPraesentationsObjekt.art'
The attribute 'XP_AbstraktesPraesentationsObjekt.art' points to an attribute of a related
feature type. Some of these attributes are replaced with a complexe attribute in XPlan 6.0.
This is the case for 'zweckbestimmung', 'detallierteZweckbestimmung', 'sondernutzung'
or 'detaillierteSondernutzung'. Therefore the mapping is changed so that the correct XPath-Syntax
is added to 'art'.
Mapping for 'art' is excluded in the structural rename functions of the sub-feature types of
'XP_AbstraktesPraesentationsObjekt'. And a lower priority is assigned to these functions, as
the ids of feature types with these new complexe attributes need to be collected first.

Also the attribute 'XP_AbstraktesPraesentationsObjekt.index' is removed with XPlan 6.0 and the
respective value should be added to 'art'. The repspective changes are done here.

SVC-1179