feat: adapt 5.4 alignments to fulfil BW-Profil requirements
In the official XPlan version the Attribut XP_Bereich.rasterBasis/XP_Rasterdarstellung is removed
in version 6.0 and marked as deprecated in version 5.4. But in the XPlan-BW-profil this change is
already mandatory in version 5.4.
To fulfil the BW-Pofil requirements the respective alignments are adapted and details given in
XP_Bereich.rasterBasis are mapped to XP_Bereich.refScan/XP_Plan.texte/XP_Plan.externeReferenz.

SVC-2069