fix: correct groovy script for 'XP_RasterdarstellungType.refText'
The syntax of the groovy script in the mapping of 'XP_RasterdarstellungType.refText' to
'XP_Plan.texte' was wrong and is corrected here.
Additonally a second instance context is added for XP_Plan.texte. Two mappings were pointing
to the same attribute at the same time without using an instance context. This led to a
wrong mapping and is corrected here.

SVC-2069